Allows the dialog to be opened when elevated. moves some helpers around
