+++ src/cascadia/TerminalSettingsModel/CascadiaSettings.idl
@@ -18,6 +18,8 @@ namespace Microsoft.Terminal.Settings.Model
 
         static String ApplicationDisplayName { get; };
         static String ApplicationVersion { get; };
+
+        static void ExportFile(String path, String content);
         
         CascadiaSettings(String userJSON, String inboxJSON);
